requirements.txt
python-decouple==3.8
googlemaps==4.10.0
googletrans==3.1.0a0
langdetect==1.0.7
Levenshtein==0.21.0
pytest==7.3.1
